fix: Ensure codegen for pitch and bearing default to empty string when matching default.
