+++ src/lib/codegen/codegen-paint.ts
@@ -1,24 +1,8 @@
+import { withDefault } from '$lib/codegen/utils';
 import { deriveColorRamp, deriveColorScale } from '$lib/interaction/color';
 import { deriveSize } from '$lib/interaction/geometry';
 import { deriveHeatmapWeight } from '$lib/interaction/weight';
 import type { CartoKitLayer } from '$lib/types';
-
-const DEFAULTS = {
-  'circle-radius': 5,
-  'fill-color': '#000000',
-  'line-color': '#000000',
-  'fill-opacity': 1,
-  'line-opacity': 1,
-  'line-width': 1,
-  'circle-color': '#000000',
-  'circle-opacity': 1,
-  'circle-stroke-color': '#000000',
-  'circle-stroke-width': 0,
-  'circle-stroke-opacity': 1,
-  'heatmap-radius': 30,
-  'heatmap-intensity': 1,
-  'heatmap-opacity': 1
-};
 
 /**
  * Generate a program fragment representing the layer's fill.
@@ -161,23 +145,4 @@ export function codegenStroke(layer: CartoKitLayer): string {
       return '';
     }
   }
-}
-
-/**
- * Return a program fragment representing a property-value pair,
- * unless the value is the same as the default value.
- *
- * @param property The property name.
- * @param cartokitValue The value of the property in the {@link CartoKitIR}.
- * @returns A (potentially empty) program fragment.
- */
-function withDefault<T extends string | number>(
-  property: keyof typeof DEFAULTS,
-  cartokitValue: T
-): string {
-  return cartokitValue !== DEFAULTS[property]
-    ? `'${property}': ${
-        typeof cartokitValue === 'string' ? `'${cartokitValue}'` : cartokitValue
-      }`
-    : '';
 }
